scatterwidgets file
