+++ tox.ini
@@ -8,6 +8,7 @@ per-file-ignores =
     src/napari_spatialdata/_constants/_constants.py: D101, D102, B024
     src/napari_spatialdata/_constants/_utils.py: D101, D102, B024
     src/napari_spatialdata/_widgets.py: D
+    src/napari_spatialdata/_scatterwidgets.py: D
 
 # D104 Missing docstring in public package
 # F401 ... imported but unused
